+++ app/src/main/AndroidManifest.xml
@@ -23,11 +23,11 @@
         <activity android:name=".ui.details.DetailsActivity"
             android:theme="@style/Theme.NoTitleActivityDialog" />
 
-        <activity android:name=".ui.CreditsActivity"
+        <activity android:name=".ui.credits.CreditsActivity"
             android:theme="@style/Theme.NoTitleActivityDialog" />
 
         <service
-            android:name=".ui.librariesview.LibrariesService"/>
+            android:name=".ui.libraries.LibrariesService"/>
     </application>
 
 </manifest>
